// src/app/modules/matenimiento/models/producto/requestVProducto.model.ts
export class RequestVProducto {
    idProducto: number=0;
    nombreProd: string ="";
    codigoProd: string ="";
    nombreUnidad: string ="";
    nombreModelo: string ="";
    precioUnitario: number=0 ;
    stock: number=0;
    estadoProducto: boolean = false;
    idUnidad: number=0;
    fotografia: string ="";
    idModelo: number=0;
    color: string ="";
    categoria: string ="";
    talla: string="";
    material: string ="";
    descripcion: string ="";
    idDetalleProducto: number=0;
}